Reorganized cpp folder into src and include
