+++ include/arm-encoder-driver/encoder_interface.h
