+++ src/encoder_interface.cpp
@@ -1,4 +1,4 @@
-#include "encoder_interface.h"
+#include "arm-encoder-driver/encoder_interface.h"
 
 int Interface::map_to_socket_can() {
     int ret;
