+++ src/main.cpp
@@ -1,4 +1,4 @@
-#include "encoder_interface.h"
+#include "arm-encoder-driver/encoder_interface.h"
 #include <boost/signals2/slot.hpp>
 #include <iostream>
